+++ src/Standards/IO/Sniffs/ControlStructures/DisallowNullCoalesceOperatorSniff.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace IsaacCodingStandard\Standards\ISAAC\Sniffs\ControlStructures;
+namespace IODigital\CodingStandard\Standards\IO\Sniffs\ControlStructures;
 
 use PHP_CodeSniffer\Files\File;
 use PHP_CodeSniffer\Sniffs\Sniff;